Calculation control parachute

+++ src/main/java/ru/vova777/coordinateParachuteNotControl/SectionTrack.java
@@ -6,19 +6,19 @@ import ru.vova777.data.SectionAltitude;
 public class SectionTrack {
 
 
-    public static int getLengthSectionTrack (SectionAltitude secAlt){
+    public static double getLengthSectionTrack (SectionAltitude secAlt){
         return secAlt.getTime() * secAlt.getWindStrength();
     }
 
-    public static int getAzimuthTrack(SectionAltitude secAlt){
+    public static double getAzimuthTrack(SectionAltitude secAlt){
         return secAlt.getAzimuthWind();
     }
 
-    public static int getFinishX(int x0, CoordinateQuarter ccq, SectionAltitude secAlt){
+    public static double getFinishX(double x0, CoordinateQuarter ccq, SectionAltitude secAlt){
         return ccq.getFinishSectionX(x0, getLengthSectionTrack(secAlt), getAzimuthTrack(secAlt));
     }
 
-    public static int getFinishY(int y0, CoordinateQuarter ccq, SectionAltitude secAlt){
+    public static double getFinishY(double y0, CoordinateQuarter ccq, SectionAltitude secAlt){
         return ccq.getFinishSectionY(y0, getLengthSectionTrack(secAlt), getAzimuthTrack(secAlt));
     }
